Move view code for publishing all assets of an volume to model.
