+++ src/zeit/content/volume/volume.py
@@ -1,3 +1,4 @@
+import itertools
 from zeit.cms.i18n import MessageFactory as _
 from zeit.cms.workflow.interfaces import IPublish
 import grokcore.component as grok
@@ -8,12 +9,15 @@ import zeit.cms.interfaces
 import zeit.cms.type
 import zeit.content.cp.interfaces
 import zeit.content.volume.interfaces
+import zeit.content.portraitbox.interfaces
+import zeit.content.infobox.interfaces
 import zeit.solr.query
 import zeit.workflow.interfaces
 import zope.interface
 import zope.lifecycleevent
 import zope.schema
 import logging
+
 
 log = logging.getLogger()
 
@@ -42,6 +46,9 @@ class Volume(zeit.cms.content.xmlsupport.XMLContentBase):
         zeit.workflow.interfaces.WORKFLOW_NS,
         'product-id')
 
+    assets_to_publish = [zeit.content.portraitbox.interfaces.IPortraitbox,
+                         zeit.content.infobox.interfaces.IInfobox
+                         ]
     @property
     def product(self):
         source = zeit.content.volume.interfaces.IVolume['product'].source(self)
@@ -203,6 +210,45 @@ class Volume(zeit.cms.content.xmlsupport.XMLContentBase):
                 log.error("Couldn't change access for {}. Skipping "
                           "it.".format(cnt.uniqueId))
         IPublish(self).publish_multiple(cnts)
+
+    def references_for_publishing(self):
+        Q = zeit.solr.query
+        additional_constraints = [
+            Q.field('published', 'not-published'),
+            Q.and_(
+                Q.bool_field('urgent', True),
+                Q.field_raw(
+                    'type',
+                    zeit.content.article.article.ArticleType.type)),
+        ]
+        articles_to_publish = self.all_content_via_solr(
+            additional_query_contstraints=additional_constraints)
+        # Flatten the list of lists and remove duplicates
+        return list(set(itertools.chain.from_iterable(
+                [self._with_references(article) for article in
+                 articles_to_publish])))
+
+    def _with_references(self, content):
+        """
+        :param content: CMSContent
+        :return: [content_ref1, content_ref2,..., content]
+        """
+        with_dependencies = [
+            content
+            for content in zeit.cms.interfaces.ICMSContentIterable(
+                content, [])
+            if self._needs_publishing(content)
+            ]
+        with_dependencies.append(content)
+        return with_dependencies
+
+    def _needs_publishing(self, content):
+        # Dont publish content which is already published
+        if zeit.cms.workflow.interfaces.IPublishInfo(content).published:
+            return False
+        # content has to provide one of interfaces defined above
+        return any([interface.providedBy(content) for interface
+                    in self.assets_to_publish])
 
 
 class VolumeType(zeit.cms.type.XMLContentTypeDeclaration):
